Handle credential error during feature store access
