+++ tests/unit/routes/test_feature.py
@@ -4,17 +4,12 @@ Tests for Feature route
 from collections import defaultdict
 from datetime import datetime
 from http import HTTPStatus
-from unittest.mock import Mock, patch
+from unittest.mock import patch
 
 import pytest
 from bson.objectid import ObjectId
 
-from featurebyte.api.feature_store import FeatureStore
-from featurebyte.core.generic import ExtendedFeatureStoreModel
-from featurebyte.exception import DocumentConflictError, DuplicatedRegistryError
-from featurebyte.feature_manager.model import ExtendedFeatureModel
-from featurebyte.models.feature_store import SourceType, SQLiteDetails, TableDetails
-from featurebyte.service.feature import FeatureService
+from featurebyte.exception import CredentialsError
 from tests.unit.routes.base import BaseApiTestSuite
 
 
@@ -182,6 +177,14 @@ class TestFeatureApi(BaseApiTestSuite):
         )
         assert feat_namespace_docs[0]["updated_at"] > feat_namespace_docs[0]["created_at"]
 
+    def test_create_401(self, test_api_client_persistent, mock_insert_feature_registry_fixture):
+        """Test create (unauthorized)"""
+        mock_insert_feature_registry_fixture.side_effect = CredentialsError
+        test_api_client, _ = test_api_client_persistent
+        self.setup_creation_route(test_api_client)
+        response = test_api_client.post(f"{self.base_route}", json=self.payload)
+        assert response.status_code == HTTPStatus.UNAUTHORIZED
+
     def test_list_404__feature_list_not_found(
         self,
         test_api_client_persistent,
@@ -289,134 +292,3 @@ class TestFeatureApi(BaseApiTestSuite):
         assert verbose_response_dict.items() > expected_info_response.items(), verbose_response.text
         assert "created_at" in verbose_response_dict
         assert verbose_response_dict["versions_info"] is not None
-
-
-@pytest.fixture(name="feature_model_dict")
-def feature_model_dict_fixture(feature_model_dict):
-    """
-    Feature model dict fixture
-    """
-    feature_model_dict["_id"] = str(ObjectId())
-    feature_model_dict["tabular_source"] = {
-        "feature_store_id": str(ObjectId()),
-        "table_details": {
-            "database_name": "sf_database",
-            "schema_name": "sf_schema",
-            "table_name": "sf_table",
-        },
-    }
-    return feature_model_dict
-
-
-@pytest.mark.asyncio
-@patch("featurebyte.session.base.BaseSession.execute_query")
-async def test_insert_feature_registry(
-    mock_execute_query,
-    feature_model_dict,
-    snowflake_connector,
-    snowflake_feature_store,
-    get_credential,
-):
-    """
-    Test insert_feature_registry
-    """
-    _ = snowflake_connector
-    feature = ExtendedFeatureModel(**feature_model_dict, feature_store=snowflake_feature_store)
-    await FeatureService(user=Mock(), persistent=Mock())._insert_feature_registry(
-        document=feature,
-        get_credential=get_credential,
-    )
-
-    match_count = 0
-    expected_partial_query = "INSERT INTO FEATURE_REGISTRY"
-    for call_args in mock_execute_query.call_args_list:
-        if expected_partial_query in call_args.args[0]:
-            match_count += 1
-    assert match_count > 0
-
-
-@pytest.fixture(name="sqlite_feature_store")
-def sqlite_feature_store_fixture(mock_get_persistent):
-    """
-    Sqlite source fixture
-    """
-    _ = mock_get_persistent
-    return FeatureStore(
-        name="sqlite_datasource",
-        type="sqlite",
-        details=SQLiteDetails(filename="some_filename"),
-    )
-
-
-@pytest.mark.asyncio
-@patch("featurebyte.session.base.BaseSession.execute_query")
-async def test_insert_feature_registry__non_snowflake_feature_store(
-    mock_execute_query, feature_model_dict, get_credential, sqlite_feature_store
-):
-    """
-    Test insert_feature_registry function (when feature store is not snowflake)
-    """
-    feature_store = ExtendedFeatureStoreModel(
-        name="sq_feature_store",
-        type=SourceType.SQLITE,
-        details=SQLiteDetails(filename="some_filename"),
-    )
-    feature_model_dict["tabular_source"] = {
-        "feature_store_id": feature_store.id,
-        "table_details": TableDetails(table_name="some_table"),
-    }
-    feature = ExtendedFeatureModel(**feature_model_dict, feature_store=sqlite_feature_store)
-    await FeatureService(user=Mock(), persistent=Mock())._insert_feature_registry(
-        document=feature, get_credential=get_credential
-    )
-    assert mock_execute_query.call_count == 0
-
-
-@pytest.mark.asyncio
-@patch("featurebyte.service.feature.FeatureManagerSnowflake")
-async def test_insert_feature_registry__duplicated_feature_registry_exception(
-    mock_feature_manager,
-    feature_model_dict,
-    get_credential,
-    snowflake_connector,
-    snowflake_feature_store,
-):
-    """
-    Test insert_feature_registry with duplicated_registry exception
-    """
-    _ = snowflake_connector
-    mock_feature_manager.return_value.insert_feature_registry.side_effect = DuplicatedRegistryError
-    feature = ExtendedFeatureModel(**feature_model_dict, feature_store=snowflake_feature_store)
-    with pytest.raises(DocumentConflictError) as exc:
-        await FeatureService(user=Mock(), persistent=Mock())._insert_feature_registry(
-            document=feature, get_credential=get_credential
-        )
-    expected_msg = (
-        'Feature (name: "sum_30m") has been registered by other feature '
-        "at Snowflake feature store."
-    )
-    assert expected_msg in str(exc)
-    assert not mock_feature_manager.return_value.remove_feature_registry.called
-
-
-@pytest.mark.asyncio
-@patch("featurebyte.service.feature.FeatureManagerSnowflake")
-async def test_insert_feature_registry__other_exception(
-    mock_feature_manager,
-    feature_model_dict,
-    get_credential,
-    snowflake_feature_store,
-    snowflake_connector,
-):
-    """
-    Test insert_feature_registry with non duplicated feature registry exception
-    """
-    _ = snowflake_connector
-    mock_feature_manager.return_value.insert_feature_registry.side_effect = ValueError
-    feature = ExtendedFeatureModel(**feature_model_dict, feature_store=snowflake_feature_store)
-    with pytest.raises(ValueError):
-        await FeatureService(user=Mock(), persistent=Mock())._insert_feature_registry(
-            document=feature,
-            get_credential=get_credential,
-        )
-    assert mock_feature_manager.return_value.remove_feature_registry.called
